# Ma.TimeManagment/OpenAPIs/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Ma.TimeManagement.Api",
    "version": "1.0"
  },
  "paths": {
    "/api/Auth/whoami": {
      "get": {
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/LoginModel"
                  }
                ]
              }
            },
            "text/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/LoginModel"
                  }
                ]
              }
            },
            "application/*+json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/LoginModel"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/Auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/RegisterModel"
                  }
                ]
              }
            },
            "text/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/RegisterModel"
                  }
                ]
              }
            },
            "application/*+json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/RegisterModel"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/Auth/save-pat": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/SavePatModel"
                  }
                ]
              }
            },
            "text/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/SavePatModel"
                  }
                ]
              }
            },
            "application/*+json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/SavePatModel"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Projects": {
      "get": {
        "tags": [
          "Projects"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TeamProjectReferenceDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TeamProjectReferenceDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TeamProjectReferenceDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Projects"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/TeamProjectReferenceAddDto"
                  }
                ]
              }
            },
            "text/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/TeamProjectReferenceAddDto"
                  }
                ]
              }
            },
            "application/*+json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/TeamProjectReferenceAddDto"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TeamProjectReferenceDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamProjectReferenceDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamProjectReferenceDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Projects/{id}": {
      "get": {
        "tags": [
          "Projects"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TeamProjectReferenceDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamProjectReferenceDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamProjectReferenceDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Projects"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/TeamProjectReferenceAddDto"
                  }
                ]
              }
            },
            "text/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/TeamProjectReferenceAddDto"
                  }
                ]
              }
            },
            "application/*+json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/TeamProjectReferenceAddDto"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Projects"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/WorkItems": {
      "get": {
        "tags": [
          "WorkItems"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkItemDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkItemDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkItemDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "WorkItems"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/WorkItemAddDto"
                  }
                ]
              }
            },
            "text/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/WorkItemAddDto"
                  }
                ]
              }
            },
            "application/*+json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/WorkItemAddDto"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/WorkItemDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkItemDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkItemDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/WorkItems/{id}": {
      "get": {
        "tags": [
          "WorkItems"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/WorkItemDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkItemDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkItemDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "WorkItems"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/WorkItemAddDto"
                  }
                ]
              }
            },
            "text/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/WorkItemAddDto"
                  }
                ]
              }
            },
            "application/*+json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/WorkItemAddDto"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": [
          "WorkItems"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/WorkItemUpdateDto"
                  }
                ]
              }
            },
            "text/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/WorkItemUpdateDto"
                  }
                ]
              }
            },
            "application/*+json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/WorkItemUpdateDto"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "WorkItems"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "EnWorkItemType": {
        "enum": [
          0,
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
      "EnWorkState": {
        "enum": [
          0,
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
      "LoginModel": {
        "type": "object",
        "properties": {
          "userID": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "LoginResult": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "nullable": true
          },
          "patConfigured": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "ProjectState": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          -2,
          -1
        ],
        "type": "integer",
        "format": "int32"
      },
      "ProjectVisibility": {
        "enum": [
          0,
          1,
          2,
          3,
          -1
        ],
        "type": "integer",
        "format": "int32"
      },
      "RegisterModel": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "nullable": true
          },
          "pat": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "RegisterResult": {
        "type": "object",
        "properties": {
          "userID": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "SavePatModel": {
        "type": "object",
        "properties": {
          "pat": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TeamProjectReferenceAddDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "defaultTeamImageUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TeamProjectReferenceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "abbreviation": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "url": {
            "type": "string",
            "nullable": true
          },
          "state": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ProjectState"
              }
            ]
          },
          "revision": {
            "type": "integer",
            "format": "int64"
          },
          "visibility": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ProjectVisibility"
              }
            ]
          },
          "defaultTeamImageUrl": {
            "type": "string",
            "nullable": true
          },
          "lastUpdateTime": {
            "type": "string",
            "format": "date-time"
          },
          "namespaceId": {
            "type": "string",
            "format": "uuid"
          },
          "requiredPermissions": {
            "type": "integer",
            "format": "int32"
          },
          "token": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "WorkItemAddDto": {
        "required": [
          "projectID"
        ],
        "type": "object",
        "properties": {
          "projectID": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "state": {
            "allOf": [
              {
                "$ref": "#/components/schemas/EnWorkState"
              }
            ]
          },
          "originalEstimate": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "remainingWork": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "completedWork": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "workItemType": {
            "allOf": [
              {
                "$ref": "#/components/schemas/EnWorkItemType"
              }
            ]
          },
          "discution": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "WorkItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "state": {
            "allOf": [
              {
                "$ref": "#/components/schemas/EnWorkState"
              }
            ]
          },
          "originalEstimate": {
            "type": "number",
            "format": "double"
          },
          "remainingWork": {
            "type": "number",
            "format": "double"
          },
          "completedWork": {
            "type": "number",
            "format": "double"
          },
          "url": {
            "type": "string",
            "nullable": true
          },
          "workItemType": {
            "allOf": [
              {
                "$ref": "#/components/schemas/EnWorkItemType"
              }
            ]
          },
          "projectID": {
            "type": "string",
            "format": "uuid"
          },
          "totalWork": {
            "type": "number",
            "format": "double",
            "readOnly": true
          },
          "projectName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "WorkItemUpdateDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "nullable": true
          },
          "state": {
            "allOf": [
              {
                "$ref": "#/components/schemas/EnWorkState"
              }
            ],
            "nullable": true
          },
          "originalEstimate": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "remainingWork": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "completedWork": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "discution": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT Authorization header. Example: <b>Bearer eyJhbGci...</b>",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
  "tags": [
    {
      "name": "Auth"
    },
    {
      "name": "Projects"
    },
    {
      "name": "WorkItems"
    }
  ]
}